Experimental AES-(128|192|256)-CFB support
In this commit, new bundled version of forge is introduced in order to
replace the old minified version, new version is easy to patch and debug.

The bundled version of forge has been patched to support stream cipher
from line 3866 to 3910. The preliminary test shows it's stable enough for use.